switch statement 2
Write a C program to implement a simple calculator. 


Description:


Get two integers (a, b) from the user. And show them 4 options (choice)



Addition

Subtraction

Multiplication

Division



if the user picks choice as 1, we need to show the addition of a, b.


if the user picks choice as 2, we need to show the subtraction of a, b.


and so on.